--- lib/Blocs/ProfileBloc/profile_event.dart
@@ -1,6 +1,0 @@
-part of 'profile_bloc.dart';
-
-@immutable
-abstract class ProfileEvent {}
-
-class GetUserProfileEvent extends ProfileEvent {}
