--- lib/Blocs/ProfileBloc/profile_state.dart
@@ -1,15 +1,0 @@
-part of 'profile_bloc.dart';
-
-@immutable
-abstract class ProfileState {}
-
-class ProfileInitial extends ProfileState {}
-
-class ProfileLoading extends ProfileState {}
-
-class ProfileLoaded extends ProfileState {
-  final String name;
-  final String email;
-
-  ProfileLoaded({required this.name, required this.email});
-}
