+++ lib/Constants/colors.dart
@@ -5,3 +5,22 @@ const Color moneyBoyPurpleLight = Color(0xFF9C91ED);
 
 //* this is a spare purple,a shade between moneyBoyPurple and moneyBoyPurpleLight
 //const Color(0xFF877BEB),
+
+const List<Color> listOfChartColors = [
+  Color(0xFF4b0082),
+  Color(0xFFa0522d),
+  Color(0xFF1e90ff),
+  Color(0xFFa52a2a),
+  Color(0xFF000080),
+  Color(0xFFbdb76b),
+  Color(0xFF2f4f4f),
+  Color(0xFF000000),
+  Color(0xFF7f0000),
+  Color(0xFF191970),
+  Color(0xFF006400),
+  Color(0xFF0000cd),
+  Color(0xFFff69b4),
+  Color(0xFFcd5c5c),
+  Color(0xFFdda0dd),
+  Color(0xFFff1493),
+];
