--- lib/Screens/profile_page.dart
@@ -1,80 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:google_fonts/google_fonts.dart';
-import 'package:moneyboi/Blocs/ProfileBloc/profile_bloc.dart';
-import 'package:moneyboi/Constants/colors.dart';
-
-class ProfilePage extends StatelessWidget {
-  const ProfilePage({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        elevation: 0.0,
-        backgroundColor: Colors.white,
-        iconTheme: const IconThemeData(color: Colors.black),
-        title: Text(
-          'PROFILE',
-          style: GoogleFonts.inter(
-            fontSize: 18.0,
-            fontWeight: FontWeight.w700,
-            color: moneyBoyPurple,
-          ),
-        ),
-      ),
-      backgroundColor: Colors.white,
-      body: SafeArea(
-        child: Column(
-          children: [
-            Center(
-              child: CircleAvatar(
-                radius: 38.0,
-                backgroundColor: Colors.grey.withOpacity(0.25),
-                child: const Icon(
-                  Icons.person,
-                  size: 32.0,
-                  color: moneyBoyPurple,
-                ),
-              ),
-            ),
-            BlocBuilder<ProfileBloc, ProfileState>(
-              builder: (context, state) {
-                if (state is ProfileLoaded) {
-                  return Column(
-                    children: [
-                      Padding(
-                        padding: const EdgeInsets.only(top: 18.0),
-                        child: Text(
-                          state.name,
-                          style: GoogleFonts.inter(
-                            fontSize: 22.0,
-                            fontWeight: FontWeight.w500,
-                            color: Colors.black.withOpacity(0.8),
-                          ),
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.only(top: 12.0),
-                        child: Text(
-                          state.email,
-                          style: GoogleFonts.inter(
-                            fontSize: 20.0,
-                            fontWeight: FontWeight.w500,
-                            color: Colors.black.withOpacity(0.8),
-                          ),
-                        ),
-                      ),
-                    ],
-                  );
-                } else {
-                  return Container();
-                }
-              },
-            )
-          ],
-        ),
-      ),
-    );
-  }
-}
