+++ lib/Screens/signup/signup_email_screen.dart
@@ -87,10 +87,12 @@ class SignupEmailWidget extends StatelessWidget {
                   ),
                   onPressed: () {
                     final regex = RegExp(
-                        r'^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$');
+                      r'^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$',
+                    );
                     if (regex.hasMatch(_emailController.text.trim())) {
                       BlocProvider.of<SignupBloc>(context).add(
-                          MoveToPasswordEvent(email: _emailController.text));
+                        MoveToPasswordEvent(email: _emailController.text),
+                      );
                     } else {}
                   },
                   child: const Icon(
