+++ lib/main.dart
@@ -1,15 +1,16 @@
 import 'package:bot_toast/bot_toast.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:get/get_navigation/src/root/get_material_app.dart';
+import 'package:get/get.dart';
 import 'package:hive/hive.dart';
 import 'package:moneyboi/Blocs/ForgotPasswordBloc/forgotpassword_bloc.dart';
-import 'package:moneyboi/Blocs/HomeScreenBloc/homescreen_bloc.dart';
-import 'package:moneyboi/Blocs/LoginBloc/login_bloc.dart';
-import 'package:moneyboi/Blocs/ProfileBloc/profile_bloc.dart';
 import 'package:moneyboi/Blocs/SignupBloc/signupbloc_bloc.dart';
-import 'package:moneyboi/Screens/home_page.dart';
-import 'package:moneyboi/Screens/login_page.dart';
+import 'package:moneyboi/Controllers/chart_screen_controller.dart';
+import 'package:moneyboi/Controllers/home_screen_controller.dart';
+import 'package:moneyboi/Controllers/login_controller.dart';
+import 'package:moneyboi/Controllers/profile_controller.dart';
+import 'package:moneyboi/Screens/home/home_page.dart';
+import 'package:moneyboi/Screens/login/login_page.dart';
 import 'package:path_provider/path_provider.dart';
 
 // ignore: avoid_void_async
@@ -28,32 +29,41 @@ class MyApp extends StatelessWidget {
     final Box _authBox = Hive.box('authBox');
     final _token = _authBox.get('token');
 
+    Get.lazyPut<LoginController>(() => LoginController(), fenix: true);
+    Get.lazyPut<HomeScreenController>(
+      () => HomeScreenController(),
+      fenix: true,
+    );
+    Get.lazyPut<ProfileController>(
+      () => ProfileController(),
+      fenix: true,
+    );
+    Get.lazyPut<ChartScreenController>(
+      () => ChartScreenController(),
+      fenix: true,
+    );
+
     return MaterialApp(
-        debugShowCheckedModeBanner: false,
-        title: 'MoneyBoi',
-        builder: BotToastInit(),
-        navigatorObservers: [BotToastNavigatorObserver()],
-        home: MultiBlocProvider(
-          providers: [
-            BlocProvider<LoginBloc>(
-              create: (context) => LoginBloc(),
-            ),
-            BlocProvider<SignupBloc>(
-              create: (context) => SignupBloc(),
-            ),
-            BlocProvider<HomeScreenBloc>(create: (context) {
-              return HomeScreenBloc();
-            }),
-            BlocProvider<ProfileBloc>(create: (context) {
-              return ProfileBloc();
-            }),
-            BlocProvider<ForgotPasswordBloc>(create: (context) {
-              return ForgotPasswordBloc();
-            }),
-          ],
-          child: GetMaterialApp(
-            home: _token != null ? const HomePage() : LoginPage(),
+      debugShowCheckedModeBanner: false,
+      title: 'MoneyBoi',
+      builder: BotToastInit(),
+      navigatorObservers: [BotToastNavigatorObserver()],
+      home: MultiBlocProvider(
+        providers: [
+          BlocProvider<SignupBloc>(
+            create: (context) => SignupBloc(),
           ),
-        ));
+          BlocProvider<ForgotPasswordBloc>(
+            create: (context) {
+              return ForgotPasswordBloc();
+            },
+          ),
+        ],
+        child: GetMaterialApp(
+          debugShowCheckedModeBanner: false,
+          home: _token != null ? const HomePage() : LoginPage(),
+        ),
+      ),
+    );
   }
 }
