+++ pubspec.yaml
@@ -30,7 +30,9 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_bloc: ^7.3.3
-  flutter_datetime_picker: ^1.5.1
+  flutter_datetime_picker:
+   git:
+    url: https://github.com/Realank/flutter_datetime_picker.git
   get: ^4.6.1
   google_fonts: ^2.1.0
   hive: ^2.0.4
